+++ whisper-service/README.md
@@ -30,10 +30,10 @@ A backend service for ScribeAR to generate transcriptions from a WAV audio strea
 
 **Implementing a new model**
 
-1. Implement `WhisperModelBase`. See `/modelBases/whisperModelBase.py` for required methods and usage.
-    * Other model bases found in `/modelBases` can be helpful for implementing commonly used functions.
-2. Create `{modelName}Requirements.txt` and populate with python dependencies for your model.
-3. Associate model key(s) to your model in `modelFactory.py`. 
+1. Implement `WhisperModelBase`. See `/modelBases/whisper_model_base.py` for required methods and usage.
+    * Other model bases found in `/model_bases` can be helpful for implementing commonly used functions.
+2. Create `{modelName}_requirements.txt` and populate with python dependencies for your model.
+3. Associate model key(s) to your model in `model_factory.py`. 
 
 **Running in development mode**
 
@@ -41,7 +41,7 @@ A backend service for ScribeAR to generate transcriptions from a WAV audio strea
 2. Configure service in `.env`
 3. Start webserver
     ```
-    fastapi dev index.py
+    fastapi dev create_server.py
     ```
 
 ### Usage
