+++ whisper-service/model_bases/whisper_model_base.py
