+++ whisper-service/model_factory.py
@@ -1,5 +1,5 @@
 from fastapi import WebSocket
-from modelBases.whisperModelBase import WhisperModelBase
+from model_bases.whisper_model_base import WhisperModelBase
 
 
 def modelFactory(modelKey: str, websocket: WebSocket) -> WhisperModelBase:
@@ -9,13 +9,13 @@ def modelFactory(modelKey: str, websocket: WebSocket) -> WhisperModelBase:
     '''
     match modelKey:
         case 'mock':
-            from models.mockWhisper import MockWhisper
+            from models.mock_whisper import MockWhisper
             return MockWhisper(websocket)
         case 'cuda-fasterwhisper-tinyen':
-            from models.fasterWhisper import FasterWhisper
+            from models.faster_whisper import FasterWhisper
             return FasterWhisper(websocket, 'tiny.en', device='cuda', localAgreeDim=2, minNewSamples=FasterWhisper.SAMPLE_RATE * 2)
         case 'cpu-fasterwhisper-tinyen':
-            from models.fasterWhisper import FasterWhisper
+            from models.faster_whisper import FasterWhisper
             return FasterWhisper(websocket, 'tiny.en', device='cpu', localAgreeDim=2, minNewSamples=FasterWhisper.SAMPLE_RATE * 2)
         case _:
             raise Exception('No model matching modelKey')
