+++ whisper-service/models/faster_whisper_requirements.txt
